Moved FetchItems and items to App.ks
