+++ inventory-app/src/ItemList.js
@@ -1,30 +1,17 @@
-import React, {useState, useEffect} from 'react';
+import React from 'react';
 // import apiURL from './api';
 
 
-export const ItemList =() => {
+export const ItemList =({items,selectedItems,setSelectedItems}) => {
 
-    const [items , setItems  ] = useState ([])
-
-    async function fetchItems (){
-        const response = await fetch ('http://localhost:3000/item')
-        const itemData = await response.json()
-        console.log(itemData)
-        setItems(itemData)
-    
-        
-    }
-
-    useEffect(()=>{
-        fetchItems()
-      }, [])
+ 
 
     return <>
 {/* maps over items, displays in divs to look pretty */}
 {items.map((anObjectMapped, index) => {
     return (
         <div className="itemCard" key={index}>
-            <img className='displayImage' src={anObjectMapped.image} />
+            <img className='displayImage' src={anObjectMapped.image} alt={anObjectMapped.title}/>
               <ul>
                 <li>{anObjectMapped.title}</li>
                 <li>£{anObjectMapped.price}</li>
